add first basic unit test
This creates an simple stab at a pattern for building testing for
os-api-ref. The basic theory is that we'll create a set of example
directories that are full doc trees with markup. We can then build
them and test the HTML output.

HTML output testing is done using BeautifulSoup so we can get
normalized HTML snippets regardless of whitespace and wrapping.

Change-Id: Idf582d62a4353db74b31183869f80bbf01b5d0f0