fix microversion parsing in parameters.yaml
This fixes the parsing of min_version and max_version in
parameters.yaml by treating things that look like floats as
strings. Otherwise 2.20 ends up folding to 2.2, which is definitely
not intentional.

This also adds a set of tests for microversion class setting in both
parameters in tables as well as in the method itself.

Change-Id: If2713fc4038e69d113cdaa7db0231a1d03f6223b