Fix Zuul jobs
The jobs are not working with Zuul v3 as is, we need
to update all job configuration and add a new job

1) Add os-api-ref-nova-src job

Replace the legacy version with a new shiny one.

2)
Avoid tox_install.sh for constraints support

Follow https://review.openstack.org/#/c/508061 to remove
tools/tox_install.sh, we can build without it.

Depends-On: I97448d681e3e48753496818fa248ee96dd36d65b
Change-Id: Ia9c8672c820a415746e43d3b1fccb2df2f574b24