Generate stable targets rather than random
A random value can still easily hit collision, and at least
one was found easily on the gates (gone with a recheck).
Moreover, the usage of a random target does not allow for ensuring
reproducible builds: the usage of an hash build from the node
content makes sure that the target is stable across different
runs of api-ref generation.

Change-Id: I3fcd8a4e5b0a66c9dbf34f4a4c472f3c93c46bb8