Mark sphinx extensions thread safe
This adds the return of some metadata to our sphinx extension setup to
indicate they are thread safe. This is needed to allow consuming
projects to do multithreaded docs builds. In some cases, this can save a
noticeable amount of time in job execution. It's absence is also a
warning in Sphinx 9.x.

Change-Id: I3a00e0836eb9dca016210d67c2f3dc3902fcec6f
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>